Fixes an issue I had with some specific images
I used a mixed approach between this library and SDWebImage. I also force the new context to be 8 bits.

I'm using this on a CocoaPods library so I can directly use this library from my fork and not the original one. If this approach works fine then I may contact the original author for a pull requrest (but now with Swift 2.0 around the corner I just needed a quick fix)